SSCS-7578 Further refactoring and tests
