+++ build.gradle
@@ -281,7 +281,7 @@ dependencies {
     compile group: 'uk.gov.hmcts.reform', name: 'doc-assembly-client', version: '1.0.6'
     compile group: 'uk.gov.hmcts.reform', name: 'document-management-client', version: '7.0.0'
 
-    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '4.0.3'
+    compile group: 'uk.gov.hmcts.reform', name: 'sscs-common', version: '4.0.4-RC-SSCS-7578'
     compile group: 'uk.gov.hmcts.reform', name: 'sscs-pdf-email-common', version: '1.4.15'
 
     testCompile group: 'io.rest-assured', name: 'rest-assured'
